refactor: move markdown utils to utils folder
